feat(app): show the account's agent email address in settings
Business owners had no in-product way to see or share the address customers
email to reach Rivus. Surface it read-only in the settings view, under
Business information, with a one-tap Copy button.

- API: return `agentEmailDomain` (from AGENT_EMAIL_DOMAIN) on the session/auth
  responses (`/v1/auth/me`, verify, accept-invite, email-change, company
  switch) so the client builds `<account.slug>@<domain>` from real config
  rather than hardcoding the domain.
- app: extract the existing CopyField (from InviteCreatedModal) into a shared
  component and reuse it for the agent address; add an "Agent email" card to
  the settings screen.
