+++ packages/api/src/http-schemas.ts
@@ -74,6 +74,13 @@ export const authResponseSchema = z
 		user: userResponseSchema,
 		account: accountResponseSchema,
 		role: roleSchema,
+		/**
+		 * Domain of the account's inbound agent address (`AGENT_EMAIL_DOMAIN`), so a
+		 * client can show/share the address customers email — `<account.slug>@<this>`
+		 * — without hardcoding the domain. Global config, not per-account data, so it
+		 * rides on the session rather than the account object.
+		 */
+		agentEmailDomain: z.string(),
 	})
 	.meta({ id: 'AuthResponse' });
 
@@ -101,6 +108,8 @@ export const sessionResponseSchema = z
 		user: userResponseSchema,
 		account: accountResponseSchema,
 		role: roleSchema,
+		/** Domain of the account's inbound agent address; see {@link authResponseSchema}. */
+		agentEmailDomain: z.string(),
 	})
 	.meta({ id: 'Session' });
