+++ packages/app/src/api/client.ts
@@ -213,6 +213,9 @@ const authResponseSchema = z.object({
 	user: userResponseSchema,
 	account: accountResponseSchema,
 	role: roleSchema,
+	// Domain of the account's inbound agent address; the UI shows the full
+	// address as `${account.slug}@${agentEmailDomain}` (see Settings).
+	agentEmailDomain: z.string(),
 });
 export type AuthResponse = z.infer<typeof authResponseSchema>;
 
@@ -228,6 +231,7 @@ const sessionResponseSchema = z.object({
 	user: userResponseSchema,
 	account: accountResponseSchema,
 	role: roleSchema,
+	agentEmailDomain: z.string(),
 });
 export type Session = z.infer<typeof sessionResponseSchema>;
